requirements.txt
black==22.12.0
click==8.1.3
Faker==16.6.1
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
six==1.16.0
